MMCA-4900-update tests and move calls for month/year display to helper formatters, correct back link flow and update platOps errors
