+++ conf/application.conf
@@ -92,10 +92,6 @@ feedback {
   source = "/CDS-FIN"
 }
 
-metrics {
-  enabled = false
-}
-
 mongodb {
   uri                 = "mongodb://localhost:27017/"${appName}
   timeToLiveInSeconds = 900
